requirements.txt
Flask==2.1.0
google-cloud-bigquery==1.22.0
gunicorn